[lesson 4] Delegate eventing logic && Implement DB logic
