+++ 04-web-framework/src/models/Eventing.ts
@@ -1,0 +1,25 @@
+type Callback = () => void;
+
+export class Eventing {
+  private events: { [key: string]: Callback[] } = {};
+
+  on(eventName: string, callback: Callback) {
+    if (this.events[eventName]) {
+      this.events[eventName].push(callback);
+    } else {
+      this.events[eventName] = [callback];
+    }
+  }
+
+  trigger(eventName: string): void {
+    const handlers = this.events[eventName];
+
+    if (!handlers || handlers.length === 0) {
+      return;
+    }
+
+    handlers.forEach(callback => {
+      callback();
+    });
+  }
+}
